requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
typing_extensions==4.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
Pillow==11.2.1
drf_yasg==1.21.10
requests==2.32.4